+++ addOns/requester/src/main/java/org/zaproxy/addon/requester/ExtensionRequester.java
@@ -26,6 +26,9 @@ import javax.swing.SwingUtilities;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
+import org.parosproxy.paros.db.Database;
+import org.parosproxy.paros.db.DatabaseException;
+import org.parosproxy.paros.db.DatabaseUnsupportedException;
 import org.parosproxy.paros.extension.ExtensionAdaptor;
 import org.parosproxy.paros.extension.ExtensionHook;
 import org.parosproxy.paros.extension.ExtensionHookMenu;
@@ -33,6 +36,8 @@ import org.parosproxy.paros.extension.ExtensionHookView;
 import org.parosproxy.paros.model.HistoryReference;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.view.View;
+import org.zaproxy.addon.requester.db.RequesterTabStorage;
+import org.zaproxy.addon.requester.db.TableRequesterTab;
 import org.zaproxy.addon.requester.internal.AbstractHttpMessageEditorDialog;
 import org.zaproxy.addon.requester.internal.ManualHttpRequestEditorPanel;
 import org.zaproxy.addon.requester.internal.RequesterOptionsPanel;
@@ -67,12 +72,15 @@ public class ExtensionRequester extends ExtensionAdaptor {
     private AbstractHttpMessageEditorDialog sendDialog;
     private AbstractHttpMessageEditorDialog resendDialog;
 
+    private final TableRequesterTab tabTable = new TableRequesterTab();
+    private final RequesterTabStorage tabStorage = new RequesterTabStorage(tabTable);
+
     public ExtensionRequester() {
         super(NAME);
         this.setOrder(211);
     }
 
-    public static final ImageIcon getManualIcon() {
+    public static ImageIcon getManualIcon() {
         if (manualIcon == null) {
             manualIcon = createIcon("hand.png");
         }
@@ -214,6 +222,17 @@ public class ExtensionRequester extends ExtensionAdaptor {
     }
 
     @Override
+    public void databaseOpen(Database db) throws DatabaseException, DatabaseUnsupportedException {
+        db.addDatabaseListener(tabTable);
+        tabTable.databaseOpen(db.getDatabaseServer());
+
+        // Reload requester panel if it is loaded
+        if (requesterPanel != null) {
+            requesterPanel.load(tabStorage);
+        }
+    }
+
+    @Override
     public boolean canUnload() {
         return true;
     }
@@ -242,4 +261,5 @@ public class ExtensionRequester extends ExtensionAdaptor {
     public String getDescription() {
         return Constant.messages.getString("requester.desc");
     }
+
 }